+++ src/const.ts
@@ -1,3 +1,0 @@
-export const DEFAULT_LOCK_ICON = 'mdi:lock-outline';
-export const DEFAULT_UNLOCK_ICON = 'mdi:lock-open-outline';
-export const DEFAULT_LOCK_DURATION = 5;